Update index.md
In this assignment, I used GitHub to code my first website. I used a combination of headers and lists to do this task. Additionally, I also re-edited the website to experiment with the website even more. Initially, I had two hashes as the first line but I wanted to see the effect of one so I had changed it to one.